requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.10.0
prophet==1.1.5
scikit-learn==1.4.0
pandas==2.2.0
pillow==10.2.0
python-dotenv==1.0.0
httpx==0.26.0
